feat: split hero section's components for better readablity
