+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Hero from "@/components/sections/hero";
+import Hero from "@/components/sections/hero/hero";
 import ContactForm from "@/components/sections/contact-form";
 import ProjectsSection from "@/components/sections/projects/projects-section";
 import Skills from "@/components/sections/skills/skills";
